[feat]: 산책 경로 생성 구현

See merge request s09-bigdata-dist-sub2/S09P22B304!20

+++ backend/src/main/java/io/ssafy/p/j9b304/backend/domain/walk/dto/request/WalkSaveRequestDto.java
@@ -13,7 +13,7 @@ public class WalkSaveRequestDto {
     Long walkId;
 
     @Schema(description = "실제 이동거리")
-    Float distance;
+    Double distance;
 
     @Schema(description = "걸음 수")
     Integer walkCount;
