+++ backend/src/main/java/io/ssafy/p/j9b304/backend/domain/walk/dto/response/WalkSaveResponseDto.java
@@ -23,7 +23,7 @@ public class WalkSaveResponseDto {
     Integer time;
 
     @Schema(description = "이동 거리(km)")
-    Float distance;
+    Double distance;
 
     @Schema(description = "걸음 수")
     Integer walkCount;
